DermAI/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0

tensorflow-macos==2.10.0
tensorflow-metal==0.5.0

numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
